Fixes
Former-commit-id: eda96a51e852d1de207286e68c6215921aee5166

+++ src/postgres/backend/utils/misc/guc.cpp
@@ -488,7 +488,7 @@ int			row_security;
 // TODO: Peloton Changes
 int			peloton_mode;
 int     peloton_layout;
-int     peloton_projectivity;
+double  peloton_projectivity;
 
 /*
  * This really belongs in pg_shmem.c, but is defined here so that it doesn't
